Changed many types to double. Required for sub pixel scrolling (needed for synchronization scrolling in the version control gui - it's easier to work with double values) . Reduces some (int) casts as well.

+++ main/src/core/Mono.Texteditor/Mono.TextEditor/CaretMoveActions.cs
@@ -295,7 +295,7 @@ namespace Mono.TextEditor
 			data.Caret.Offset = data.Document.Length;
 		}
 		
-		public static int LineHeight { get; set; }
+		public static double LineHeight { get; set; }
 		
 		public static void PageUp (TextEditorData data)
 		{
